Updating python code.
Tinker -g now prints (but does not save) a complete XML specification. This
specification has not been tested in Quartus, or in the TCL Scripts.

Several TODOs:
Sharing of OCT pins still needs to be done by hand (it should and can be automatic)
The structure of XML files has changed (fmax, fref, etc now declared in a Phy)